+++ CHANGELOG.md
@@ -1,12 +1,17 @@
-# 0.0.2+1
+# [0.1.0]
+
+- Migrate to null-safety
+- Change `EJStepState.active` to `EJStepState.enable`
+
+## [0.0.2+1]
 
 - Updated README.md
 
-## 0.0.2
+## [0.0.2]
 
 - Updated README.md
 - Added Example
 
-## 0.0.1
+## [0.0.1]
 
 - Initial Release